refactor: 컨트롤러 패키지 구조 개선 - 도메인 별 하위 패키지 구성
