+++ src/main/java/webserver/controller/ControllerResolver.java
@@ -1,5 +1,7 @@
 package webserver.controller;
 
+import webserver.controller.file.FileController;
+import webserver.controller.user.UserSaveController;
 import webserver.utils.HttpMethod;
 
 import java.util.HashMap;
