+++ src/test/java/webserver/controller/file/FileControllerTest.java
@@ -1,9 +1,10 @@
-package webserver.controller;
+package webserver.controller.file;
 
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.controller.file.FileController;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
 import webserver.http.HttpStatus;
