+++ src/test/java/webserver/controller/user/UserSaveControllerTest.java
@@ -1,4 +1,4 @@
-package webserver.controller;
+package webserver.controller.user;
 
 import db.Database;
 import model.User;
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.controller.user.UserSaveController;
 import webserver.http.HttpRequest;
 import webserver.http.HttpResponse;
 import webserver.http.HttpStatus;
